modify the source tree for chisel project
