+++ hw/chisel/src/main/scala/snax/csr_namaner/CsrManagerTestParameters.scala
@@ -1,0 +1,12 @@
+package snax
+
+package csr_manager
+
+import chisel3._
+import chisel3.util._
+
+object csrManagerTestParameters {
+  def csrNum = 20
+  def csrAddrWidth = log2Up(csrNum)
+  def csrModuleTagName = "Test"
+}
